config and utils functionality seperated - ha
