+++ config/db.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose');
+
+//connect to mongodb server
+const connectDB = async () => {
+    try {
+        await mongoose.connect(process.env.MONGO_URI || 'mongodb://localhost:27017/csvdata', {
+            useNewUrlParser: true,
+            useUnifiedTopology: true,
+        });
+        console.log('MongoDB Connected...');
+    } catch (err) {
+        console.error('Error connecting to MongoDB:', err.message);
+        process.exit(1);
+    }
+};
+
+module.exports = connectDB;
