+++ routes/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const dynamicRoute = require('./dynamicRoute');
+
+const router = express.Router();
+
+router.use('/dynamic', dynamicRoute);
+
+module.exports = router;
